+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/measure/ActionMeter.kt
@@ -6,13 +6,11 @@ import com.atlassian.performance.tools.jiraactions.api.ActionResult
 import com.atlassian.performance.tools.jiraactions.api.ActionType
 import com.atlassian.performance.tools.jiraactions.api.measure.output.ActionMetricOutput
 import com.atlassian.performance.tools.jiraactions.api.measure.output.ThrowawayActionMetricOutput
-import com.atlassian.performance.tools.jiraactions.api.w3c.DisabledW3cPerformanceTimeline
 import com.atlassian.performance.tools.jiraactions.api.w3c.RecordedPerformanceEntries
 import com.atlassian.performance.tools.jiraactions.api.w3c.W3cPerformanceTimeline
 import java.time.Clock
 import java.time.Duration
 import java.util.*
-import java.util.function.Predicate
 import javax.json.JsonObject
 
 /**
@@ -21,15 +19,11 @@ import javax.json.JsonObject
  * @param virtualUser identifies a virtual user
  * @param output writes the metrics
  * @param clock measures latencies
- * @param w3cPerformanceTimeline records W3C performance entries
- * @param drilldownCondition condition based on ActionMetric upon which drilldown is performed
  */
 class ActionMeter private constructor(
     private val virtualUser: UUID,
     private val output: ActionMetricOutput,
     private val clock: Clock,
-    private val w3cPerformanceTimeline: W3cPerformanceTimeline,
-    private val drilldownCondition: Predicate<ActionMetric>,
     private val postMetricHook: PostMetricHook
 ) {
 
@@ -40,7 +34,7 @@ class ActionMeter private constructor(
             "\noutput = output" +
             "\n)" +
             "\n.clock(clock)" +
-            "\n.performanceTimeline(w3cPerformanceTimeline)" +
+            "\n.postMetricHook(DrillDownHook(w3cPerformanceTimeline))" +
             "\n.build()")
     )
     constructor(
@@ -52,9 +46,7 @@ class ActionMeter private constructor(
         virtualUser = virtualUser,
         output = output,
         clock = clock,
-        w3cPerformanceTimeline = w3cPerformanceTimeline,
-        drilldownCondition = Predicate { true },
-        postMetricHook = NoopPostMetricHook()
+        postMetricHook = DrillDownHook(w3cPerformanceTimeline)
     )
 
     @Deprecated(
@@ -74,8 +66,6 @@ class ActionMeter private constructor(
         virtualUser = virtualUser,
         output = output,
         clock = clock,
-        w3cPerformanceTimeline = DisabledW3cPerformanceTimeline(),
-        drilldownCondition = Predicate { true },
         postMetricHook = NoopPostMetricHook()
     )
 
@@ -123,7 +113,6 @@ class ActionMeter private constructor(
             ).virtualUser(virtualUser)
             postMetricHook.run(actionMetricBuilder)
             result?.let { actionMetricBuilder.observation(observation(result)) }
-            conditionallyAddDrilldown(actionMetricBuilder)
             output.write(actionMetricBuilder.build())
             return result
         } catch (e: Exception) {
@@ -140,21 +129,8 @@ class ActionMeter private constructor(
                 result = actionResult
             ).virtualUser(virtualUser)
             postMetricHook.run(actionMetricBuilder)
-            conditionallyAddDrilldown(actionMetricBuilder)
             output.write(actionMetricBuilder.build())
             throw Exception("Action '${key.label}' $actionResult", e)
-        }
-    }
-
-    private fun conditionallyAddDrilldown(actionMetricBuilder: ActionMetric.Builder): ActionMetric {
-        return actionMetricBuilder.build().let {
-            if (drilldownCondition.test(it)) {
-                actionMetricBuilder
-                    .drilldown(w3cPerformanceTimeline.record())
-                    .build()
-            } else {
-                it
-            }
         }
     }
 
@@ -165,8 +141,6 @@ class ActionMeter private constructor(
         virtualUser = virtualUser,
         output = output,
         clock = clock,
-        w3cPerformanceTimeline = w3cPerformanceTimeline,
-        drilldownCondition = Predicate { true },
         postMetricHook = NoopPostMetricHook()
     )
 
@@ -174,8 +148,6 @@ class ActionMeter private constructor(
         private val output: ActionMetricOutput
     ) {
         private var virtualUser: UUID = UUID.randomUUID()
-        private var w3cPerformanceTimeline: W3cPerformanceTimeline = DisabledW3cPerformanceTimeline()
-        private var drilldownCondition: Predicate<ActionMetric> = Predicate { true }
         private var clock = Clock.systemUTC()
         private var postMetricHook: PostMetricHook = NoopPostMetricHook()
 
@@ -183,24 +155,17 @@ class ActionMeter private constructor(
             meter.output
         ) {
             clock = meter.clock
-            w3cPerformanceTimeline = meter.w3cPerformanceTimeline
-            drilldownCondition = meter.drilldownCondition
             virtualUser = meter.virtualUser
         }
 
-        fun performanceTimeline(w3cPerformanceTimeline: W3cPerformanceTimeline) = apply { this.w3cPerformanceTimeline = w3cPerformanceTimeline }
-        fun drilldownCondition(drilldownCondition: Predicate<ActionMetric>) = apply { this.drilldownCondition = drilldownCondition }
         fun clock(clock: Clock) = apply { this.clock = clock }
         fun postMetricHook(postMetricHook: PostMetricHook) = apply { this.postMetricHook = postMetricHook }
-
         fun virtualUser(virtualUser: UUID) = apply { this.virtualUser = virtualUser }
 
         fun build(): ActionMeter = ActionMeter(
             virtualUser = virtualUser,
             output = output,
             clock = clock,
-            w3cPerformanceTimeline = w3cPerformanceTimeline,
-            drilldownCondition = drilldownCondition,
             postMetricHook = postMetricHook
         )
     }